requirements.txt
# Core ETL Dependencies
boto3==1.42.29
botocore==1.42.29
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
jmespath==1.0.1
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
s3transfer==0.16.0
six==1.17.0
urllib3==2.6.3

# Data Analysis & Visualization
pandas>=2.0.0
matplotlib>=3.7.0
plotly>=5.14.0
seaborn>=0.12.0

# Forecasting & Statistics
statsmodels>=0.14.0
prophet>=1.1.0
numpy>=1.24.0
scipy>=1.10.0
